Us2.3 aimer et commenter les publications

+++ src/main/java/miage/groupe6/reseausocial/model/jpa/service/ActionPostService.java
@@ -31,7 +31,7 @@ public class ActionPostService {
         }
     }
 
-    public long countLikes(Post post) {
+    public int countLikes(Post post) {
         return apr.countByPostAndStatut(post, StatutActionPost.LIKE);
     }
 }
